test for czech

+++ jointtrainer_multilang.py
@@ -720,7 +720,7 @@ class JointTrainer:
                 max_len = self.nertrainreader
 
             print("Eval interval is set to {} ".format(self.args['eval_interval']))
-        self.args["epochs"] = 5 * max_len / self.args["eval_interval"]
+        self.args["epochs"] = int(5 * max_len / self.args["eval_interval"])
         self.args['ner_cats'] = len(self.nertrainreader.label_voc)
         print("Word vocabulary size  : {}".format(len(self.nertrainreader.word_voc)))
         self.args['vocab'] = self.whole_vocab.w2ind
